video 16

+++ src/main/java/com/lothin/phoneshp/dto/PaginationDTO.java
@@ -6,11 +6,11 @@ import lombok.Data;
 @Data
 @Builder
 public class PaginationDTO {
-    private int numberOfElements;
-    private int number;
-    private int size;
-    private int totalElements;
-    private int totalPages;
+    private Long numberOfElements;
+    private Long number;
+    private Long size;
+    private Long totalElements;
+    private Long totalPages;
     private boolean empty;
     private boolean first;
     private boolean last;
